Enable on-type formatting from language server
